% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_glmnet.R
\name{shiny_glmnet}
\alias{shiny_glmnet}
\title{Shiny application for glmnet}
\usage{
shiny_glmnet(data = NULL, family = "gaussian")
}
\arguments{
\item{data}{list, containing y and x matrices to run glmnet on, Default: NULL}

\item{family}{character, Response type}
}
\description{
Interactive application that runs the glmnet package functions
}
\details{
if data is null then sample data is used
}
\examples{
\dontrun{
if(interactive()){
shiny_glmnet()
 }
}
}
